+++ app/services/newton_api/positions.rb
@@ -2,7 +2,9 @@ module NewtonApi
   class Positions < NewtonApi::Sync
     def self.update(integration)
       res = HTTParty.get("https://api.newton.co/v1/balances", headers: headers(integration))
-      integration.positions.destroy_all
+      if integration.positions.present?
+        integration.positions.destroy_all
+      end
       res.each do |symbol, value|
         next unless value != 0.0
         integration.positions << Cryptocurrency.new(
