+++ app/services/questrade_api/balances.rb
@@ -2,7 +2,9 @@ module QuestradeApi
   class Balances < QuestradeApi::Sync
     def self.update(integration, account)
       res = HTTParty.get("https://api01.iq.questrade.com/v1/accounts/#{account["number"]}/balances", headers: headers(integration))
-      integration.balances.destroy_all
+      if integration.balances.present?
+        integration.balances.destroy_all
+      end
       res["perCurrencyBalances"].each do |b|
         integration.balances.create(
           currency: b["currency"],
